refactor: define varaible dbnameDict in a separate file MA-908
